LayoutTests/fast/scrolling/ios/touch-scroll-visibility-hidden-expected.txt

PASS hitElement.id is "under-target"
PASS scroller.scrollTop is 0
swipe complete
PASS scroller.scrollTop is 0
PASS document.scrollingElement.scrollTop is 90
PASS successfullyParsed is true

TEST COMPLETE
